# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "FKMed",
    "description" : "FKMed — health plan beneficiary portal API",
    "version" : "0.12.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/reimbursement-previews" : {
      "get" : {
        "tags" : [ "reimbursement-preview-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReimbursementPreviewListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "reimbursement-preview-controller" ],
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "request" : {
                    "$ref" : "#/components/schemas/ReimbursementPreviewRequest"
                  },
                  "documents" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "request" ]
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementPreviewResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/recovery/request" : {
      "post" : {
        "tags" : [ "recovery-controller" ],
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordRecoveryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        }
      }
    },
...
    "/api/tokens/current" : {
      "get" : {
        "tags" : [ "token-controller" ],
        "operationId" : "current",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reimbursements/{id}" : {
      "get" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementDetailView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reimbursement-previews/{id}" : {
      "get" : {
        "tags" : [ "reimbursement-preview-controller" ],
        "operationId" : "detail_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementPreviewResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/legal-documents/{type}" : {
      "get" : {
        "tags" : [ "legal-document-controller" ],
        "operationId" : "document",
        "parameters" : [ {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LegalDocumentView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/context/beneficiaries/{beneficiaryId}" : {
      "get" : {
        "tags" : [ "context-controller" ],
        "operationId" : "beneficiary",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BeneficiarySummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TokenView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ReimbursementDocumentRequest" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string",
            "enum" : [ "RECEIPT", "MEDICAL_ORDER", "COMPLEMENTARY", "BUDGET" ]
          },
          "fileName" : {
            "type" : "string"
          }
        },
        "required" : [ "category" ]
      },
...
      "ReimbursementPreviewRequest" : {
        "type" : "object",
        "properties" : {
          "beneficiaryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "expenseTypeCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReimbursementDocumentRequest"
            }
          }
        },
        "required" : [ "beneficiaryId", "expenseTypeCode" ]
      },
      "ReimbursementPreviewResult" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "expenseType" : {
            "type" : "string"
          },
          "beneficiary" : {
            "type" : "string"
          },
          "situation" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "CONCLUIDA" ]
          },
          "estimatedValue" : {
            "type" : "number"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "concludedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "base" : {
            "type" : "string"
          },
          "disclaimer" : {
            "type" : "string"
          }
        }
      },
...
      "PasswordRecoveryRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          }
        },
        "required" : [ "email" ]
      },
...
      "BankView" : {
        "type" : "object",
        "properties" : {
          "bankCode" : {
            "type" : "string"
          },
          "agency" : {
            "type" : "string"
          },
          "account" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "CORRENTE", "POUPANCA" ]
          }
        }
      },
      "DocumentView" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string",
            "enum" : [ "RECEIPT", "MEDICAL_ORDER", "COMPLEMENTARY", "BUDGET" ]
          },
          "fileName" : {
            "type" : "string"
          },
          "fileSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "uploadedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "GlosaView" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
      "PendencyView" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "deadlineAt" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ProviderView" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "councilCode" : {
            "type" : "string"
          },
          "councilNumber" : {
            "type" : "string"
          },
          "councilUf" : {
            "type" : "string"
          },
          "specialty" : {
            "type" : "string"
          }
        }
      },
      "ReimbursementDetailView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "expenseType" : {
            "type" : "string"
          },
          "beneficiary" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          },
          "careDate" : {
            "type" : "string",
            "format" : "date"
          },
          "requestedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expectedPaymentDate" : {
            "type" : "string",
            "format" : "date"
          },
          "amountRequested" : {
            "type" : "number"
          },
          "amountReimbursed" : {
            "type" : "number"
          },
          "glosa" : {
            "$ref" : "#/components/schemas/GlosaView"
          },
          "denialReason" : {
            "type" : "string"
          },
          "pendency" : {
            "$ref" : "#/components/schemas/PendencyView"
          },
          "bank" : {
            "$ref" : "#/components/schemas/BankView"
          },
          "provider" : {
            "$ref" : "#/components/schemas/ProviderView"
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DocumentView"
            }
          },
          "timeline" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimelineView"
            }
          },
          "regulatoryNote" : {
            "type" : "string"
          }
        }
      },
      "TimelineView" : {
        "type" : "object",
        "properties" : {
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          },
          "description" : {
            "type" : "string"
          }
        }
      },
...
      "ReimbursementPreviewListItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "expenseType" : {
            "type" : "string"
          },
          "beneficiary" : {
            "type" : "string"
          },
          "requestedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "situation" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "CONCLUIDA" ]
          },
          "estimatedValue" : {
            "type" : "number"
          }
        }
      },
...
      "LegalDocumentView" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          },
          "publishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "body" : {
            "type" : "string"
          },
          "acceptedByMe" : {
            "type" : "boolean"
          }
        }
      },
...
      "BeneficiarySummary" : {
        "type" : "object",
        "properties" : {
          "beneficiaryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "firstName" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "TITULAR", "DEPENDENT" ]
          },
          "planName" : {
            "type" : "string"
          },
          "cardNumber" : {
            "type" : "string"
          },
          "avatarUrl" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}